+++ src/common/mod.rs
@@ -4,10 +4,10 @@ mod id;
 mod immutable;
 mod mutable;
 mod node;
-mod response;
+mod routing_table;
 
 pub use id::*;
 pub use immutable::*;
 pub use mutable::*;
 pub use node::*;
-pub use response::*;
+pub use routing_table::*;
