+++ src/rpc/server/mod.rs
@@ -1,0 +1,7 @@
+//! modules needed only for nodes running in server mode.
+
+pub mod peers;
+pub mod tokens;
+
+pub use peers::*;
+pub use tokens::*;
