+++ src/rpc/socket.rs
